pool-scheduler-project/requirements.txt
black~=22.8
celery~=5.2
coverage~=6.4
crispy-bootstrap5~=0.7
debugpy~=1.6.3
django-auto-prefetch~=1.3.0
django-coverage-plugin~=2.0
django-crispy-forms~=1.14
django-debug-toolbar~=3.7
django-extensions~=3.2
django-range-merge~=0.2
django-render-block~=0.9.2
Django~=4.1
factory-boy~=3.2
Faker~=15.0
flake8-isort~=4.2
flake8~=5.0
gunicorn~=20.1.0
isort[requirements_deprecated_finder]~=5.10
psycopg2-binary~=2.9
pytest-cov~=3.0
pytest~=7.1
redis~=4.3
werkzeug~=2.2

